Moved WalletStateAlgebra into the DataApi package
